Add center -> [direction] -> center gestures
When I type password I have to switch between capital and small letters.
This is a hassle with 8vim. To switch to the case I have to lift my
finger -- as that adds a space -- I have to delete it and then select
the shit key. Adding the gesture center -> top -> center should fix
this. The hope is also that it helps increase the flow while writing
capitalized words: German has a lot of them -- not just at the start of
a sentence.

When I mistype, it would be nice to fix it and continue. However having
to lift the finger and remove the additional space seems like many
unnecessary steps. So I added the gesture center -> top -> center.

The other two directions I added to be able to correct myself when I
head into the wrong sector for starting a letter. They don't do
anything.